% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleEnroll.R
\name{enroll_uiA}
\alias{enroll_uiA}
\alias{enroll_uiB}
\alias{enroll_serverA}
\alias{enroll_serverB}
\title{Enrollment Module}
\usage{
enroll_uiA(id)

enroll_uiB(id)

enroll_serverA(id, site, data)

enroll_serverB(id, site, data)
}
\arguments{
\item{id}{id for bundled module UI and module Server}

\item{site}{study site (country)}

\item{data}{engineered data extracted from synapse database}
}
\description{
UI and Server for Enrollment Module
}
