% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{box}
\alias{box}
\alias{dashboard_title}
\alias{tab_enroll_output}
\title{UI Elements}
\usage{
box(...)

dashboard_title(sideBar_width)

tab_enroll_output()
}
\arguments{
\item{...}{UI elements for the box}

\item{sideBar_width}{(numeric) width of sidebar}

\item{id}{id for bundled module UI and module Server}

\item{site}{study site (country)}

\item{data}{engineered data extracted from synapse database}
}
\description{
Elements for UI.
}
